refactor(releases): share release matrix loader
