+++ apps/web/tests/unit/queries/prefetch-dashboard.test.ts
@@ -11,7 +11,7 @@ const mocks = vi.hoisted(() => ({
   ]),
 }));
 
-vi.mock('@/app/app/(shell)/dashboard/releases/actions', () => ({
+vi.mock('@/app/app/(shell)/dashboard/releases/release-matrix-loader', () => ({
   loadReleaseMatrix: mocks.loadReleaseMatrix,
 }));
